♻️ Move tenant-scope types to lib to break a feature cycle
scheduled-event importing SpaceContentScope from features/space closed a
cycle: space imports scheduled-event for its loaders, and user reaches it
too. The scope types are cross-cutting contracts consumed by every
content feature, so they now live in lib (importable by all, imports
nothing); features/space/types re-exports them so existing import sites
keep the documented path.
